requirements.txt
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
greenlet==3.2.1
idna==3.10
pillow==11.2.1
pyaes==1.6.1
pyasn1==0.6.1
pyTelegramBotAPI==4.26.0
requests==2.32.3
rsa==4.9.1
SQLAlchemy==2.0.40
telebot==0.0.5
Telethon==1.40.0
typing_extensions==4.13.2
urllib3==2.4.0
